requirements.txt
python-telegram-bot
google-generativeai
matplotlib
requests